// src/app/services/notifications/models/notification-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationResponse {
  channel?: 'EMAIL' | 'WEBSOCKET' | 'PUSH' | 'SMS';
  createdAt?: string;
  id?: number;
  message?: string;
  metadata?: {
[key: string]: string;
};
  projectId?: number;
  readAt?: string;
  sentAt?: string;
  status?: 'PENDING' | 'SENT' | 'DELIVERED' | 'READ' | 'FAILED';
  taskId?: number;
  title?: string;
  type?: 'TASK_ASSIGNED' | 'TASK_UPDATED' | 'TASK_COMPLETED' | 'PROJECT_INVITATION' | 'PROJECT_UPDATED' | 'DEADLINE_REMINDER' | 'COMMENT_ADDED' | 'MENTION' | 'SYSTEM_ANNOUNCEMENT';
  userId?: number;
}